+++ client/src/api/mod.rs
@@ -8,10 +8,6 @@ mod block_builder;
 mod consolidation;
 mod types;
 
-pub use self::{
-    address::*,
-    block_builder::{pow::*, *},
-    types::*,
-};
+pub use self::{address::*, block_builder::*, types::*};
 
 const ADDRESS_GAP_RANGE: u32 = 20;