+++ pow/src/miner.rs
@@ -1,7 +1,7 @@
-// Copyright 2020-2021 IOTA Stiftung
+// Copyright 2020-2022 IOTA Stiftung
 // SPDX-License-Identifier: Apache-2.0
 
-//! Contains a nonce provider that mine nonces.
+//! Multi-threaded PoW miner.
 
 use std::{
     sync::{
@@ -12,7 +12,7 @@ use std::{
 };
 
 use crypto::{
-    encoding::ternary::{b1t6, Btrit, T1B1Buf, TritBuf},
+    encoding::ternary::{b1t6, T1B1Buf, TritBuf},
     hashes::{
         blake2b::Blake2b256,
         ternary::{
@@ -22,37 +22,22 @@ use crypto::{
         Digest,
     },
 };
-use thiserror::Error;
 
-use crate::providers::{NonceProvider, NonceProviderBuilder};
+use crate::{score::count_trailing_zeros, Error, LN_3};
 
 const DEFAULT_NUM_WORKERS: usize = 1;
-// Precomputed natural logarithm of 3 for performance reasons.
-// See https://oeis.org/A002391.
-const LN_3: f64 = 1.098_612_288_668_109;
 
-/// Errors occurring when computing nonces with the `Miner` nonce provider.
-#[derive(Error, Debug)]
-pub enum Error {
-    /// The worker has been cancelled.
-    #[error("the worker has been cancelled")]
-    Cancelled,
-    /// Invalid proof of work score.
-    #[error("invalid proof of work score {0}, requiring {1} trailing zeros")]
-    InvalidPowScore(u32, usize),
-}
-
-/// A type to cancel the `Miner` nonce provider to abort operations.
+/// A type to cancel a [`Miner`] to abort operations.
 #[derive(Default, Clone)]
 pub struct MinerCancel(Arc<AtomicBool>);
 
 impl MinerCancel {
-    /// Creates a new `MinerCancel`.
+    /// Creates a new [`MinerCancel`].
     pub fn new() -> Self {
         Self::default()
     }
 
-    /// Cancels the `Miner` nonce provider.
+    /// Cancels the [`Miner`].
     pub fn trigger(&self) {
         self.0.store(true, Ordering::Relaxed);
     }
@@ -62,13 +47,13 @@ impl MinerCancel {
         self.0.load(Ordering::Relaxed)
     }
 
-    /// Reset the cancel flag.
+    /// Resets the cancel flag.
     fn reset(&self) {
         self.0.store(false, Ordering::Relaxed);
     }
 }
 
-/// Builder for the `Miner` nonce provider.
+/// Builder for a [`Miner`].
 #[derive(Default)]
 #[must_use]
 pub struct MinerBuilder {
@@ -77,23 +62,25 @@ pub struct MinerBuilder {
 }
 
 impl MinerBuilder {
-    /// Sets the desired number of workers for the `Miner` nonce provider.
+    /// Creates a new [`MinerBuilder`].
+    pub fn new() -> Self {
+        Self { ..Default::default() }
+    }
+
+    /// Sets the desired number of workers for the [`Miner`].
     pub fn with_num_workers(mut self, num_workers: usize) -> Self {
         self.num_workers.replace(num_workers);
         self
     }
 
-    /// Sets a `MinerCancel to abort the `Miner` nonce provider.
+    /// Sets a `MinerCancel to abort the [`Miner`].
     pub fn with_cancel(mut self, cancel: MinerCancel) -> Self {
         self.cancel.replace(cancel);
         self
     }
-}
 
-impl NonceProviderBuilder for MinerBuilder {
-    type Provider = Miner;
-
-    fn finish(self) -> Miner {
+    /// Builds the [`Miner`].
+    pub fn finish(self) -> Miner {
         Miner {
             num_workers: self.num_workers.unwrap_or(DEFAULT_NUM_WORKERS),
             cancel: self.cancel.unwrap_or_else(MinerCancel::new),
@@ -101,7 +88,7 @@ impl NonceProviderBuilder for MinerBuilder {
     }
 }
 
-/// A nonce provider that mine nonces.
+/// A multi-threaded pow nonce miner.
 pub struct Miner {
     num_workers: usize,
     cancel: MinerCancel,
@@ -132,9 +119,7 @@ impl Miner {
             }
 
             for (i, hash) in hasher.hash().enumerate() {
-                let trailing_zeros = hash.iter().rev().take_while(|t| *t == Btrit::Zero).count();
-
-                if trailing_zeros >= target_zeros {
+                if count_trailing_zeros(&hash) >= target_zeros {
                     cancel.trigger();
                     return Ok(nonce + i as u64);
                 }
@@ -145,13 +130,9 @@ impl Miner {
 
         Err(Error::Cancelled)
     }
-}
 
-impl NonceProvider for Miner {
-    type Builder = MinerBuilder;
-    type Error = Error;
-
-    fn nonce(&self, bytes: &[u8], target_score: u32) -> Result<u64, Self::Error> {
+    /// Mines a nonce for provided bytes.
+    pub fn nonce(&self, bytes: &[u8], target_score: u32) -> Result<u64, Error> {
         self.cancel.reset();
 
         let mut nonce = 0;
@@ -160,7 +141,7 @@ impl NonceProvider for Miner {
             (((bytes.len() + std::mem::size_of::<u64>()) as f64 * target_score as f64).ln() / LN_3).ceil() as usize;
 
         if target_zeros > HASH_LENGTH {
-            return Err(Self::Error::InvalidPowScore(target_score, target_zeros));
+            return Err(Error::InvalidPowScore(target_score, target_zeros));
         }
 
         let worker_width = u64::MAX / self.num_workers as u64;
@@ -188,4 +169,21 @@ impl NonceProvider for Miner {
 
         Ok(nonce)
     }
+}
+
+fn _get_miner(bytes: &[u8], min_pow_score: u32, num_workers: usize) -> Result<u64, Error> {
+    MinerBuilder::new()
+        .with_num_workers(num_workers)
+        .finish()
+        .nonce(bytes, min_pow_score)
+}
+
+/// Returns a closure for a miner with `num_cpus` workers.
+pub fn get_miner(min_pow_score: u32) -> impl Fn(&[u8]) -> Result<u64, Error> {
+    move |bytes| _get_miner(bytes, min_pow_score, num_cpus::get())
+}
+
+/// Returns a closure for a miner with `num_workers` workers.
+pub fn get_miner_num_workers(min_pow_score: u32, num_workers: usize) -> impl Fn(&[u8]) -> Result<u64, Error> {
+    move |bytes| _get_miner(bytes, min_pow_score, num_workers)
 }